requirements.txt
flask==2.3.3
openai==1.3.5
tavily-python==0.3.0
trafilatura==1.6.4
pypdf==3.17.1
requests==2.31.0
python-dotenv==1.0.0
ollama==0.1.7